Larger buffers for speedup
The original algorithm did a lot of buffer resize/copy operations, as
git feeds data in 64k chunks. This ends up being an O(n^2) algorithm,
which for a 200MB data file is a real killer, making a git diff take 4
minutes. This change resizes the buffer in 10M blocks, which cuts it
down to 5 seconds. Much more acceptable! ;)

Long term it would be nice if it was not necessary to keep the whole
file in memory, but given that there is no way to figure out the
filename/hash before the whole file is piped, there is no easy way to do
that.