Added support for pageant, ssh authentication agent on Windows.
git-jam will default to try to connect using pageant if
process.platform is 'win32', SSH_AUTH_SOCK is not specified and
private key cannot be found in ~/.ssh/id_rsa.

For better support it should probably be configurable, because
with this change, trying to connect without password and not
configured keys will result in "Authentication failure." error
from ssh instead of "Please set up as password..." error from
git-jam.